+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v0.1 - October 1? 2024
 - Gameinfo is inside an Enum Class inside of the `game_util_gameinfo.py` file
 - Main Game loop is inside the `game_main.py` file
+- Error Handler exists in the `game_util_error.py` file